In the print template we'll need to be able to show a barcode for the Cassandra row  of the just-created record.
Ideally this would be shown as a small barcode so it can just be
scanned, but with the text ID underneath. This should solve any problems
with finding the right database record associated with the membership
request.

There shouldn't be any privacy concerns in this as we would only expose
to an external attacker which row IDs we assigned, but they're 128-bit
random GUIDs. It should be hard to guess them, and even if they're
guessed they don't expose any information about the requestor.