Add functionality for users to download their membership agreement and vcard file.
This adds takeout compliance as requested by the GDPR guidelines. All
data stored about a member is displayed in the overview page, and the
member can download their agreement PDF (which is a different format and
thus not in the overview) and download their own contact details as a
VCF file (standard format export requirement).

Deletion and correction are not (currently) handled by the form, instead
they refer people to contact a person.