Update Member Creator to use the DB API directly.
It was using the Cassandra Thrift API before, which has been deprecated
a while ago; this also meant that it would not benefit immediately from
any improvements to db.go. So the API was updated to accomodate the
missing functions from the member creator.

Change in functionality: instead of creating all members in the database
at the same time, the database records will be created one by one.